fix(fixing user role):adding user roles

+++ src/resolvers/loginUserResolver.ts
@@ -1,5 +1,6 @@
 import { AuthenticationError } from "apollo-server";
 import { LoggedUserModel } from "../models/AuthUser";
+import adminEmail from "../database/db.config";
 
 export const loggedUserResolvers: any = {
   Query: {
@@ -9,9 +10,10 @@ export const loggedUserResolvers: any = {
       return upvalue;
     },
     async getUsers_Logged(_: any, args: any, ctx: any, amount: any) {
-      // if (!ctx.currentUser) {
-      //   throw new AuthenticationError("You must be logged in");
-      // }
+      console.log(ctx)
+      if (!ctx.currentUser || ctx.currentUser.role !== 'admin') {
+        throw new AuthenticationError("Only administrators can access this.");
+      }
       return await LoggedUserModel.find().sort({ createdAt: -1 }).limit(amount);
     },
   },
@@ -20,23 +22,31 @@ export const loggedUserResolvers: any = {
       _: any,
       { userInput: { name, email, picture } }: any
     ) {
+      const isAdmin = email == adminEmail;
       const createdUser = new LoggedUserModel({
         name,
         email,
         picture,
+        role:isAdmin?"admin" : "applicant",
         createdAt: new Date().toISOString(),
       });
 
       const res = await createdUser.save(); // MongoDB saving
       return res;
     },
-    async deleteUser_Logged(_: any, { ID }: any) {
+    async deleteUser_Logged(_: any, { ID }: any,  ctx: any) {
+      if (!ctx.currentUser || ctx.currentUser.role !== 'admin') {
+        throw new AuthenticationError("Only administrators can delete userss.");
+      }
       const wasDeleted = (await LoggedUserModel.deleteOne({ _id: ID }))
         .deletedCount;
       return wasDeleted; //1 if something was deleted, 0 if nothing deleted
     },
 
-    async updateUser_Logged(_: any, { ID, editUserInput: { name } }: any) {
+    async updateUser_Logged(_: any, { ID, editUserInput: { name } }: any, ctx: any) {
+      // if (!ctx.currentUser || ctx.currentUser.role !== 'admin') {
+      //   throw new AuthenticationError("Only administrators can update user.");
+      // }
       const wasEdited = (await LoggedUserModel.updateOne({ _id: ID }, { name }))
         .modifiedCount;
       return wasEdited; //1||true if something was Edited, 0||true if nothing Edited
